Resolve unlink() error
Resolves unlink() errors that occur when concurrent requests for a cache file results in one unlink() call being successful while other unlink() calls result in a PHP error being thrown.

+++ core/xpdo/cache/xpdocachemanager.class.php
@@ -511,10 +511,12 @@ class xPDOCacheManager {
                             }
                             elseif (is_file($path)) {
                                 if (is_array($extensions) && !empty($extensions) && !$this->endsWith($file, $extensions)) continue;
-                                if (unlink($path)) {
-                                    array_push($result, $path);
-                                } else {
-                                    $hasMore= true;
+                                if(file_exists($path)){
+                                    if (unlink($path)) {
+                                        array_push($result, $path);
+                                    } else {
+                                        $hasMore= true;
+                                    }
                                 }
                             }
                         }
